requirements.txt
blake3
numpy
Pillow
requests
